Make location text more specific
Edited placeholder text in location section. When we tested the app with
a user, there were issues with inputting a postcode rather then a full
address. To get accurate results, a full address is needed. This text
will encourage users to enter that information. We also added an
ellipses to be consistent in style with the search by school name
section.
- 'home address, suburb or postcode' -> 'please enter your home
address...'